move old test files around to make room for improved testing and better folder structure
