+++ package.json
@@ -5,10 +5,10 @@
   "dependencies": {
     "@auth0/auth0-react": "^1.6.0",
     "@craco/craco": "^6.2.0",
-    "autoprefixer": "^10.4.1",
+    "autoprefixer": "^9",
     "axios": "^0.21.1",
     "cra-template": "1.1.2",
-    "postcss": "^8.4.5",
+    "postcss": "^7",
     "prop-types": "^15.8.0",
     "react": "^17.0.2",
     "react-confetti": "^6.0.1",
@@ -18,7 +18,7 @@
     "react-scripts": "4.0.3",
     "redux": "^4.1.1",
     "redux-thunk": "^2.3.0",
-    "tailwindcss": "^3.0.8"
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.1",
