+++ src/pages/Game/Game.test.js
@@ -1,10 +1,11 @@
-import { render, screen } from '../utils';
-import Game from '../../pages/Game';
+import { render, screen } from '../../utils';
 import { 
     CHOICE_GAME_PATH,
     SOCIAL_TYPE,
     GAME_NOT_AVAILABLE_TEMP_MESSAGE
 } from '../../const';
+
+import Game from './';
 
 describe('Game', () => {
     test('renders message if game not available', () => {
