+++ src/pages/Game/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
 
-import WhatIsThisGame from '../components/Games/WhatIsThis/WhatIsThisGame';
+import WhatIsThisGame from '../../components/Games/WhatIsThis/WhatIsThisGame';
 
 import {
     DEFAULT,
@@ -9,7 +9,7 @@ import {
     ANIMAL_TYPE,
     FOOD_TYPE,
     GAME_NOT_AVAILABLE_TEMP_MESSAGE
-} from '../const';
+} from '../../const';
 
 // note: temp function created while I'm still developing rest of the games
 const canRenderGame = (wordType = DEFAULT.STRING) => {
